\encoding{UTF-8}
\name{arabidopsisE}
\alias{arabidopsisE}
\title{Gene expression data for Arabidopsis thaliana}
\description{
This data set has been extracted from the R package "NBPSeq"; see the listed reference.
There are around 22000 genes, each of which is under two conditions. For each gene under
each condition, there are 3 replicates. The discrete count for each replicate
is modelled by negative binomial distribution, and the exact negative binomial
test is used to test if a gene is differentially expressed.
}
\references{ 
 Di, Y., Schafer, D. W., Cumbie, J. S. and Chang, J. H. (2011). The NBP negative
 binomial model for assessing differential gene expression from RNA-Seq, Statistical
 Applications in Genetics and Molecular Biology 10(1): 24 pages.
}
