\encoding{UTF-8}
\name{hivdata}
\alias{hivdata}
\title{HIV data}
\description{
This data set has been analyzed and provided by the listed reference. There are
118 positions, each of which is under two types of HIV. Each type has 73 subjects.
}
\references{ 
 Gilbert, P. B. (2005). A modified false discovery rate multiple-comparisons procedure for discrete
data, applied to human immunodeficiency virus genetics, J. R. Statist. Soc. Ser. C 54(1): 143-158.
}